Wrapped up managercontroller authentication

+++ src/main/java/com/revature/project02/controllers/ManagerController.java
@@ -52,6 +52,7 @@ public class ManagerController {
 	 * @throws JsonMappingException 
 	 * @throws JsonParseException 
 	 * @throws - BadRequestException - if the manager parameter could not be instantiated. Usually by bad json.
+	 * @throws - UnauthorizedException
 	 */
 	@PostMapping
 	public ResponseEntity<Manager> addManager(@RequestParam("manager") String managerStr, @RequestParam("password") String password, @RequestHeader("token") String token) {
@@ -82,29 +83,48 @@ public class ManagerController {
 	public ResponseEntity<Manager> getDriverManager(@PathVariable("id") Integer id, @RequestHeader("token") String token){
 		UnencryptedAuthenticationToken uat = AuthTokenUtil.fromEncryptedAuthenticationToken(token);
 		if(uat == null) throw new UnauthorizedException("Unauthorized Access!");
-		if(!"class com.revature.project02.models.Admin".equals(uat.getRole())) throw new UnauthorizedException("Unauthorized Access!");
+		if("class com.revature.project02.models.Driver".equals(uat.getRole()))
+		{
+			if (uat.getUserId().equals(id)) return new ResponseEntity<>(mService.getManagerByDriver(dService.getDriverById(id)), HttpStatus.OK);
+			else throw new UnauthorizedException("Unauthorized Access!");
+		}
+		else if("class com.revature.project02.models.Manager".equals(uat.getRole())) return new ResponseEntity<>(mService.getManagerByDriver(dService.getDriverById(id)), HttpStatus.OK);
+		else if("class com.revature.project02.models.Admin".equals(uat.getRole())) return new ResponseEntity<>(mService.getManagerByDriver(dService.getDriverById(id)), HttpStatus.OK);
+		else throw new UnauthorizedException("Unauthorized Access!");
 		
-		return new ResponseEntity<>(mService.getManagerByDriver(dService.getDriverById(id)), HttpStatus.OK);
 	}
 	
 	/**
 	 * Description - Returns a List of all managers in the db table
+	 * @param token - String representing the encrypted authentication token
 	 * @return - a List of all managers in the db table
 	 * @throws - ResourceNotFoundException - if there are no managers in the table
 	 */
 	@GetMapping
-	public ResponseEntity<List<Manager>> getAllManagers(){
+	public ResponseEntity<List<Manager>> getAllManagers(@RequestHeader("token") String token){
+		//Unencrypt the token
+		UnencryptedAuthenticationToken uat = AuthTokenUtil.fromEncryptedAuthenticationToken(token);
+		if(uat == null) throw new UnauthorizedException("Unauthorized Access!");
+		if(!"class com.revature.project02.models.Admin".equals(uat.getRole())) throw new UnauthorizedException("Unauthorized Access!");
+
 		return new ResponseEntity<>(mService.getAllManagers(), HttpStatus.OK);
 	}
 	
 	/**
 	 * Description - Deletes the requested manager
 	 * @param id - Integer representation of the manager's id
+	 * @param token - String representing the encrypted authentication token
 	 * @return - HttpStatus of ok
 	 * @throws - BadRequestException
+	 * @throws UnauthorizedException
 	 */
 	@DeleteMapping(value = "/managerid-{id}")
-	public HttpStatus deleteManager(@PathVariable("id") Integer id) {
+	public HttpStatus deleteManager(@PathVariable("id") Integer id, @RequestHeader("token") String token) {
+		//Unencrypt the token
+		UnencryptedAuthenticationToken uat = AuthTokenUtil.fromEncryptedAuthenticationToken(token);
+		if(uat == null) throw new UnauthorizedException("Unauthorized Access!");
+		if(!"class com.revature.project02.models.Admin".equals(uat.getRole())) throw new UnauthorizedException("Unauthorized Access!");
+
 		mService.deleteManager(id);
 		return HttpStatus.OK;
 	}
